Pull some emscripten stuff into its own files to reduce clutter in main
